requirements.txt
Appium-Python-Client==2.11.1
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
joblib==1.3.2
langdetect==1.0.9
Naked==0.1.32
nltk==3.8.1
outcome==1.2.0
packaging==23.1
Pillow==10.0.0
pycparser==2.21
PySimpleGUI==4.60.5
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
selenium==4.10.0
selenium-stealth==1.0.6
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
Terminalia==0.0.4
textblob==0.17.1
tls-client==0.2.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
webdriver-manager==4.0.0
wsproto==1.2.0
